+++ docs/changelog.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [UNRELEASED]
+
+## [0.15.0] - 2025-06-09
 ### Added
 - add support for python 3.13
 - add support for django 5.2